pyproject.toml
[tool.poetry]
name = "ScopeSim"
version = "0.9.2a0"
description = "Generalised telescope observation simulator"
license = "GPL-3.0-or-later"
authors = ["Kieran Leschinski <kieran.leschinski@unive.ac.at>"]
maintainers = [
    "Kieran Leschinski <kieran.leschinski@unive.ac.at>",
    "Hugo Buddelmeijer <hugo@buddelmeijer.nl>",
    "Fabian Haberhauer <fabian.haberhauer@univie.ac.at>",
]
readme = "README.md"
repository = "https://github.com/AstarVienna/ScopeSim"
documentation = "https://scopesim.readthedocs.io/en/latest/"
classifiers = [
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Astronomy",
]

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
numpy = "^1.26.3"
scipy = "^1.14.1"
astropy = "^6.0.1"
matplotlib = "^3.8.2"
pooch = "^1.8.2"

docutils = "^0.19"
httpx = ">=0.26.0"
beautifulsoup4 = "^4.12.1"
lxml = "^5.2.2"
pyyaml = "^6.0.1"
more-itertools = "^10.2.0"
tqdm = "^4.66.5"

synphot = "^1.4.0"
skycalc-ipy = ">=0.5.1"
anisocado = ">=0.3.0"
astar-utils = ">=0.3.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0"
jupytext = "^1.10.0"
ipykernel = "^6.24.0"
ipympl = "^0.9.4"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
scopesim_templates = ">=0.6.0"
ipykernel = "^6.24.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = "^5.3.0"
sphinx-book-theme = "^1.1.0"
jupyter-sphinx = "^0.2.3"
sphinx-copybutton = "^0.5.2"
myst-nb = "^1.0.0"
sphinxcontrib-apidoc = "^0.4.0"
nbsphinx = "^0.9.3"
numpydoc = "^1.6.0"
scopesim_templates = ">=0.6.0"
ipykernel = "^6.24.0"
ipympl = "^0.9.4"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/AstarVienna/ScopeSim/issues"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--strict-markers"
markers = [
    "webtest: marks tests as requiring network (deselect with '-m \"not webtest\"')",
    "slow: marks tests that take a while (deselect with '-m \"not slow\"')",
]
filterwarnings = [
    "error",
    # Should probably be fixed:
    "ignore::ResourceWarning",
    "ignore:The fit may be poorly conditioned:astropy.utils.exceptions.AstropyUserWarning",
    "ignore::astropy.units.core.UnitsWarning",
    # Perhaps fix?
    "ignore:divide by zero encountered in scalar divide:RuntimeWarning",
    "ignore:divide by zero encountered in double_scalars:RuntimeWarning",
    "ignore:invalid value encountered in multiply:RuntimeWarning",
    "ignore:invalid value encountered in divide:RuntimeWarning",
    "ignore:Cannot merge meta key.*:astropy.utils.metadata.MergeConflictWarning",
    "ignore::astropy.wcs.wcs.FITSFixedWarning",
    "default:The fov_grid*:DeprecationWarning",
    # Raised when saving fits files, not so important to fix:
    "ignore:.*a HIERARCH card will be created.*:astropy.io.fits.verify.VerifyWarning",
    # Web-related issues, fix at some point
    "ignore:'urllib3.contrib.pyopenssl'*:DeprecationWarning",
    "ignore:Blowfish*:UserWarning",
    "ignore:TripleDES*:UserWarning",
    # Not sure what that is but it's everywhere...
    "ignore:'cgi'*:DeprecationWarning",
    "ignore:The py23*:DeprecationWarning",
    "ignore:datetime.datetime.utcfromtimestamp()*:DeprecationWarning",
    # Pytest Notebook stuff
    "ignore:Proactor*:RuntimeWarning",
    # lxml warning introduced by beautifulsoup
    "ignore:The 'strip_cdata' option of HTMLParser*:DeprecationWarning",
]

[tool.coverage.report]
omit = ["scopesim/tests/*", "docs/*"]
